+++ Util/BuildTools/Setup.sh
@@ -157,7 +157,7 @@ for PY_VERSION in ${PY_VERSION_LIST[@]} ; do
 
     pushd ${BOOST_BASENAME}-source >/dev/null
 
-    BOOST_TOOLSET="clang-$CARLA_LLVM_VERSION_MAJOR.0"
+    BOOST_TOOLSET="clang-$CARLA_LLVM_VERSION_MAJOR.1.0"
     BOOST_CFLAGS="-fPIC -std=c++14 -DBOOST_ERROR_CODE_HEADER_ONLY"
 
     py3="/usr/bin/env python${PY_VERSION}"